make 'start' stop the remote instance before trying to start it again
For when stop wouldn't reach the server, and then 'start' would be run
again with the daemon on the server still running.

Just to make it "clean"; it seemed to work before already, mostly?
Suspecting that it might have been when it would go into 100% CPU
eating mode, although no, that was when using stoprouting and
startrouting. Anyway?